+++ test/helper.rb
@@ -18,6 +18,12 @@ def testpath(path)
   File.join(TEST_DIR, path)
 end
 
+def commit_details
+  { :message => "Did something at #{Time.now}",
+    :name => "Tom Preston-Werner",
+    :email => "tom@github.com" }
+end
+
 # test/spec/mini 3
 # http://gist.github.com/25455
 # chris@ozmm.org
@@ -26,7 +32,7 @@ def context(*args, &block)
   return super unless (name = args.first) && block
   require 'test/unit'
   klass = Class.new(defined?(ActiveSupport::TestCase) ? ActiveSupport::TestCase : Test::Unit::TestCase) do
-    def self.test(name, &block) 
+    def self.test(name, &block)
       define_method("test_#{name.gsub(/\W/,'_')}", &block) if block
     end
     def self.xtest(*args) end