+++ test/test_file.rb
@@ -11,9 +11,12 @@ context "File" do
   end
 
   test "existing file" do
-    file = @wiki.file("Mordor/todo.txt")
+    commit = @wiki.repo.commits.first
+    file   = @wiki.file("Mordor/todo.txt")
     assert_equal "[ ] Write section on Ents\n", file.raw_data
-    assert_equal @wiki.repo.commits.first.id, file.version.id
+    assert_equal 'todo.txt',         file.name
+    assert_equal commit.id,          file.version.id
+    assert_equal commit.author.name, file.version.author.name
   end
 
   test "accessing tree" do