+++ test/test_markup.rb
@@ -6,10 +6,6 @@ context "Markup" do
     FileUtils.rm_rf(@path)
     Grit::Repo.init_bare(@path)
     @wiki = Gollum::Wiki.new(@path)
-
-    @commit = { :message => "Add stuff",
-                :name => "Tom Preston-Werner",
-                :email => "tom@github.com" }
   end
 
   teardown do
@@ -17,12 +13,18 @@ context "Markup" do
   end
 
   test "formats page from Wiki#pages" do
-    @wiki.write_page("Bilbo Baggins", :markdown, "a [[Foo]][[Bar]] b", @commit)
+    @wiki.write_page("Bilbo Baggins", :markdown, "a [[Foo]][[Bar]] b", commit_details)
     assert @wiki.pages[0].formatted_data
   end
 
+  #########################################################################
+  #
+  # Links
+  #
+  #########################################################################
+
   test "double page links no space" do
-    @wiki.write_page("Bilbo Baggins", :markdown, "a [[Foo]][[Bar]] b", @commit)
+    @wiki.write_page("Bilbo Baggins", :markdown, "a [[Foo]][[Bar]] b", commit_details)
 
     # "<p>a <a class=\"internal absent\" href=\"/Foo\">Foo</a><a class=\"internal absent\" href=\"/Bar\">Bar</a> b</p>"
     page    = @wiki.page("Bilbo Baggins")
@@ -41,7 +43,7 @@ context "Markup" do
   end
 
   test "double page links with space" do
-    @wiki.write_page("Bilbo Baggins", :markdown, "a [[Foo]] [[Bar]] b", @commit)
+    @wiki.write_page("Bilbo Baggins", :markdown, "a [[Foo]] [[Bar]] b", commit_details)
 
     # "<p>a <a class=\"internal absent\" href=\"/Foo\">Foo</a> <a class=\"internal absent\" href=\"/Bar\">Bar</a> b</p>"
     page = @wiki.page("Bilbo Baggins")
@@ -60,7 +62,7 @@ context "Markup" do
   end
 
   test "page link" do
-    @wiki.write_page("Bilbo Baggins", :markdown, "a [[Bilbo Baggins]] b", @commit)
+    @wiki.write_page("Bilbo Baggins", :markdown, "a [[Bilbo Baggins]] b", commit_details)
 
     page = @wiki.page("Bilbo Baggins")
     output = page.formatted_data
@@ -70,7 +72,7 @@ context "Markup" do
   end
 
   test "absent page link" do
-    @wiki.write_page("Tolkien", :markdown, "a [[J. R. R. Tolkien]]'s b", @commit)
+    @wiki.write_page("Tolkien", :markdown, "a [[J. R. R. Tolkien]]'s b", commit_details)
 
     page = @wiki.page("Tolkien")
     output = page.formatted_data
@@ -83,7 +85,7 @@ context "Markup" do
     ["/wiki", "/wiki/"].each_with_index do |path, i|
       name = "Bilbo Baggins #{i}"
       @wiki = Gollum::Wiki.new(@path, :base_path => path)
-      @wiki.write_page(name, :markdown, "a [[#{name}]] b", @commit)
+      @wiki.write_page(name, :markdown, "a [[#{name}]] b", commit_details)
 
       page = @wiki.page(name)
       output = page.formatted_data
@@ -94,16 +96,22 @@ context "Markup" do
   end
 
   test "external page link" do
-    @wiki.write_page("Bilbo Baggins", :markdown, "a [[http://example.com]] b", @commit)
+    @wiki.write_page("Bilbo Baggins", :markdown, "a [[http://example.com]] b", commit_details)
 
     page = @wiki.page("Bilbo Baggins")
     assert_equal "<p>a <a href=\"http://example.com\">http://example.com</a> b</p>", page.formatted_data
   end
 
+  #########################################################################
+  #
+  # Images
+  #
+  #########################################################################
+
   test "image with http url" do
     ['http', 'https'].each do |scheme|
       name = "Bilbo Baggins #{scheme}"
-      @wiki.write_page(name, :markdown, "a [[#{scheme}://example.com/bilbo.jpg]] b", @commit)
+      @wiki.write_page(name, :markdown, "a [[#{scheme}://example.com/bilbo.jpg]] b", commit_details)
 
       page = @wiki.page(name)
       output = page.formatted_data
@@ -116,7 +124,7 @@ context "Markup" do
     index = @wiki.repo.index
     index.add("alpha.jpg", "hi")
     index.commit("Add alpha.jpg")
-    @wiki.write_page("Bilbo Baggins", :markdown, "a [[/alpha.jpg]] [[a | /alpha.jpg]] b", @commit)
+    @wiki.write_page("Bilbo Baggins", :markdown, "a [[/alpha.jpg]] [[a | /alpha.jpg]] b", commit_details)
 
     page = @wiki.page("Bilbo Baggins")
     assert_equal %{<p>a <img src="/wiki/alpha.jpg" /><a href="/wiki/alpha.jpg">a</a> b</p>}, page.formatted_data
@@ -217,11 +225,17 @@ context "Markup" do
     relative_image(content, output)
   end
 
+  #########################################################################
+  #
+  # File links
+  #
+  #########################################################################
+
   test "file link with absolute path" do
     index = @wiki.repo.index
     index.add("alpha.jpg", "hi")
     index.commit("Add alpha.jpg")
-    @wiki.write_page("Bilbo Baggins", :markdown, "a [[Alpha|/alpha.jpg]] b", @commit)
+    @wiki.write_page("Bilbo Baggins", :markdown, "a [[Alpha|/alpha.jpg]] b", commit_details)
 
     page = @wiki.page("Bilbo Baggins")
     output = Gollum::Markup.new(page).render
@@ -247,6 +261,12 @@ context "Markup" do
     page = @wiki.page("Bilbo Baggins")
     assert_equal %{<p>a <a href="http://example.com/alpha.jpg">Alpha</a> b</p>}, page.formatted_data
   end
+
+  #########################################################################
+  #
+  # Code
+  #
+  #########################################################################
 
   test "code blocks" do
     content = "a\n\n```ruby\nx = 1\n```\n\nb"
@@ -296,6 +316,12 @@ context "Markup" do
     compare(content, output)
   end
 
+  #########################################################################
+  #
+  # Various
+  #
+  #########################################################################
+
   test "escaped wiki link" do
     content = "a '[[Foo]], b"
     output = "<p>a [[Foo]], b</p>"
@@ -317,17 +343,29 @@ context "Markup" do
     compare(content, output, 'org')
   end
 
-  test "tex block syntax" do
+  #########################################################################
+  #
+  # TeX
+  #
+  #########################################################################
+
+  test "TeX block syntax" do
     content = 'a \[ a^2 \] b'
     output = "<p>a <script type=\"math/tex; mode=display\">a^2</script> b</p>"
     compare(content, output, 'md')
   end
 
-  test "tex inline syntax" do
+  test "TeX inline syntax" do
     content = 'a \( a^2 \) b'
     output = "<p>a <script type=\"math/tex\">a^2</script> b</p>"
     compare(content, output, 'md')
   end
+
+  #########################################################################
+  #
+  # Helpers
+  #
+  #########################################################################
 
   def compare(content, output, ext = "md", regexes = [])
     index = @wiki.repo.index
@@ -349,6 +387,7 @@ context "Markup" do
     index.add("greek/Bilbo-Baggins.md", content)
     index.commit("Add alpha.jpg")
 
+    @wiki.clear_cache
     page = @wiki.page("Bilbo Baggins")
     rendered = Gollum::Markup.new(page).render
     assert_equal output, rendered