+++ HISTORY.md
@@ -3,6 +3,14 @@
 * Major Enhancements
   * Optimize page write/update/delete to use Grit::Index::read_tree instead
     of manually recreating entire index contents.
+  * Added --irb option for the gollum command.
+  * Update working dir (if present) when edited via the API (#6)
+* Minor Enhancements
+  * Support a `:gollum_path` Sinatra setting for `Precious::App`
+  * Add Wiki#size to efficiently count pages without loading them.
+  * Add the correct content type when serving files from the frontend.
+  * Wiki#pages come back sorted by Page#title.
+  * Add --host option and default it to 127.0.0.1.
 * Bug Fixes
   * Increase minimum Sanitize version requirement to 1.1.0.
     1.0.x versions of Sanitize require Hpricot instead of Nokogiri
@@ -13,6 +21,7 @@
     anonymous info.
   * Prevent `Gollum::Wiki#write_page` from clobbering existing pages.
   * Handle duplicate page errors in frontend.
+  * Fix bugs trying to retrieve pages with invalid names.
 
 # 1.0.1 / 2010-08-12